refactor(dashboard): extract leaf components from approvals megapage
Pull three leaf components out of `src/app/approvals/page.tsx` into
`src/app/approvals/_components/`:

- CountdownTimer  — inline timer for approval expiry
- Spinner         — inline loading indicator (kept distinct from
                    `patchwork/Spinner`; different visual / a11y)
- RiskMeter       — risk-level bar gauge (uses `patchwork/RiskPill`)

Page shrinks from 1492 → 1396 lines (-96). The `EmptyState` and
`Spinner` defined inline are visually distinct from their generic
counterparts in `src/components/patchwork/` (different roles and
styling), so they are kept page-local rather than collapsed.

The `ApprovalCard`, `BatchActionBar`, and approvals-specific
`EmptyState` remain inline — they have heavier coupling to page state
and props and are out of scope for this PR.

Verified:
- npx tsc --noEmit → clean
- npm test → 24/24 pass
