+++ dashboard/src/app/approvals/_components/Spinner.tsx
@@ -1,0 +1,17 @@
+export function Spinner() {
+  return (
+    <span
+      style={{
+        display: "inline-block",
+        width: 14,
+        height: 14,
+        border: "2px solid currentColor",
+        borderTopColor: "transparent",
+        borderRadius: "50%",
+        animation: "spin 0.6s linear infinite",
+        verticalAlign: "middle",
+      }}
+      aria-hidden="true"
+    />
+  );
+}
